+++ components/login/ButtonFindAccount.tsx
@@ -1,5 +1,4 @@
 import Link from 'next/link'
-import { useRouter } from 'next/router'
 import React from 'react'
 import { useSetRecoilState } from 'recoil'
 import { findAccountSelector } from 'state/atoms'
